+++ tests/unit/capi/GEOSPreparedDistanceWithinTest.cpp
@@ -58,6 +58,7 @@ group test_capigeosprepareddistancewithin_group("capi::GEOSPreparedDistanceWithi
 // Test Cases
 //
 
+// Empty inputs
 template<>
 template<>
 void object::test<1>
@@ -71,6 +72,7 @@ void object::test<1>
     );
 }
 
+// Disjoint polygons
 template<>
 template<>
 void object::test<2>
@@ -85,6 +87,7 @@ void object::test<2>
 
 }
 
+// Point contained in polygon
 template<>
 template<>
 void object::test<3>
@@ -98,6 +101,7 @@ void object::test<3>
     );
 }
 
+// Disjoint line and point
 template<>
 template<>
 void object::test<4>
@@ -111,6 +115,7 @@ void object::test<4>
     );
 }
 
+// Intersecting lines
 template<>
 template<>
 void object::test<5>
@@ -124,6 +129,7 @@ void object::test<5>
     );
 }
 
+// Intersecting polygon and line
 template<>
 template<>
 void object::test<6>
@@ -137,6 +143,7 @@ void object::test<6>
     );
 }
 
+// Empty geometries
 template<>
 template<>
 void object::test<7>
@@ -150,6 +157,7 @@ void object::test<7>
     );
 }
 
+// Empty geometries
 template<>
 template<>
 void object::test<8>
@@ -163,6 +171,7 @@ void object::test<8>
     );
 }
 
+// Mixed empty and non-empty
 template<>
 template<>
 void object::test<9>
@@ -176,6 +185,7 @@ void object::test<9>
     );
 }
 
+// Mixed empty and non-empty
 template<>
 template<>
 void object::test<10>
@@ -186,6 +196,20 @@ void object::test<10>
         "POLYGON EMPTY",
         geos::DoubleInfinity,
         0
+    );
+}
+
+// Prepared geometry contained in test geometry
+template<>
+template<>
+void object::test<11>
+()
+{
+    checkDistanceWithin(
+        "POLYGON((1 1,1 5,5 5,5 1,1 1))",
+        "POLYGON((0 0, 0 10, 10 10, 10 0, 0 0))",
+        0,
+        1
     );
 }