+++ NEWS.md
@@ -18,7 +18,8 @@ xxxx-xx-xx
   - GeometryPrecisionReducer: Return correct dimensionality for empty results (GH-684, Dan Baston)
   - Improve performance of coverage union (GH-681, Dan Baston)
   - Improve performance of prepared polygon intersection (GH-690, Dan Baston)
-  - Implement indexed calculations for prepared geometry distance (GH-691, Dan Baston)
+  - Improve performance of prepared polygon distance (GH-693, Dan Baston)
+  - Implement indexed calculations for prepared geometry isWithinDistance (GH-691, Dan Baston)
 
 
 ## Changes in 3.11.0